Rename packages to be plural.
